+++ src/Service/Athena/src/Enum/DataCatalogType.php
@@ -4,7 +4,6 @@ namespace AsyncAws\Athena\Enum;
 
 final class DataCatalogType
 {
-    public const FEDERATED = 'FEDERATED';
     public const GLUE = 'GLUE';
     public const HIVE = 'HIVE';
     public const LAMBDA = 'LAMBDA';
@@ -12,7 +11,6 @@ final class DataCatalogType
     public static function exists(string $value): bool
     {
         return isset([
-            self::FEDERATED => true,
             self::GLUE => true,
             self::HIVE => true,
             self::LAMBDA => true,